Moving a few files to utilities from detail
