+++ examples/2d_plot/2d_plot.cpp
@@ -4,7 +4,7 @@
 #include <goofit/PDFs/combine/ProdPdf.h>
 #include <goofit/UnbinnedDataSet.h>
 #include <goofit/Variable.h>
-#include <goofit/detail/Style.h>
+#include <goofit/utilities/Style.h>
 
 #include <RVersion.h>
 #include <TCanvas.h>
