+++ include/goofit/utilities/Style.h
